Save and load - all parameters
+Game score accounted

+++ Assets/_Wizards/1_Scripts/ILevelInfo.cs
@@ -9,6 +9,6 @@ namespace WizardsPlatformer
         LevelConfig GetLevelConfig();
 
         void AccountForBonuses(Dictionary<BonusType, int> bonuses);
-        void AccountForScore(int levelScore);
+        void AccountForScore(float levelScore);
     }
 }
